// frontend/src/app/core/api/model/projectCreationDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProjectCreationDto { 
    name: string;
    projectKey: string;
    teamId: number;
    startDate?: string;
    endDate?: string;
    budget?: number;
}